Improve time preferences components' readability - Break up time-preferences file into multiple files for better readability - Make minor improvements to page performance with memoization
